requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
aws-psycopg2==1.2.1
boto3==1.17.25
botocore==1.20.25
colorama==0.4.4
importlib-metadata==3.7.0
iniconfig==1.1.1
jmespath==0.10.0
more-itertools==8.7.0
numpy==1.20.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
zipp==3.4.0
